blog_backend/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-filter==21.1
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
importlib-metadata==4.11.4
lorem==0.1.1
Markdown==3.3.7
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
